// core/models/grading.ts
export const NONE = 'NONE';
export const SWISS_QUARTER_GRADES = 'SWISS_QUARTER_GRADES';
export const SWISS_HALF_GRADES = 'SWISS_HALF_GRADES';
export const SWISS_TENTH_GRADES = 'SWISS_TENTH_GRADES';
export const UZH_MEF_HALF_GRADES = 'UZH_MEF_HALF_GRADES';
export const PASS_FAIL = 'PASS_FAIL';
export const PARTICIPATION = 'PARTICIPATION';
export const PERCENTAGE = 'PERCENTAGE';
export const PASS_CONDITIONAL_FAIL = 'PASS_CONDITIONAL_FAIL';
export const NOT_GRADED = 'NOT_GRADED';
export const UNKNOWN = 'UNKNOWN';
export const HUNDRED_POINT_SCALE = 'HUNDRED_POINT_SCALE';

export type Grading =
	| 'NONE'
	| 'SWISS_QUARTER_GRADES'
	| 'SWISS_HALF_GRADES'
	| 'SWISS_TENTH_GRADES'
	| 'UZH_MEF_HALF_GRADES'
	| 'PASS_FAIL'
	| 'PARTICIPATION'
	| 'PERCENTAGE'
	| 'PASS_CONDITIONAL_FAIL'
	| 'NOT_GRADED'
	| 'HUNDRED_POINT_SCALE'
	| 'UNKNOWN';